refactor: invite to workspace modal [#132]
