+++ src/lib/components/header/global-search-bar/GlobalSearchBar.svelte
